@@ -1,49 +1,50 @@
 <?php
 
-
-
 namespace App\Exceptions;
 
-use Throwable;
 use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
+use Throwable;
 
 class Handler extends ExceptionHandler
 {
     /**
+     * A list of exception types with their corresponding custom log levels.
+     *
+     * @var array<class-string<\Throwable>, \Psr\Log\LogLevel::*>
+     */
+    protected $levels = [
+        //
+    ];
+
+    /**
      * A list of the exception types that are not reported.
      *
-     * @var array
+     * @var array<int, class-string<\Throwable>>
      */
     protected $dontReport = [
         //
     ];
 
     /**
-     * A list of the inputs that are never flashed for validation exceptions.
+     * A list of the inputs that are never flashed to the session on validation exceptions.
      *
-     * @var array
+     * @var array<int, string>
      */
     protected $dontFlash = [
+        'current_password',
         'password',
         'password_confirmation',
     ];
 
+    /**
+     * Register the exception handling callbacks for the application.
+     *
+     * @return void
+     */
     public function register()
     {
         $this->reportable(function (Throwable $e) {
             //
         });
     }
-    // public function report(Throwable $exception)
-    // {
-    // }
-    // public function shouldReport(Throwable $exception)
-    // {
-    // }
-    // public function render($request, Throwable $exception)
-    // {
-    // }
-    // public function renderForConsole($output, Throwable $exception)
-    // {
-    // }
 }
