@@ -2,12 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Event;
-use App\Sport;
-use App\Student;
-use App\User;
-use Illuminate\Http\Request;
-use Mockery\Undefined;
+use App\Models\Event;
 use phpDocumentor\Reflection\PseudoTypes\LowercaseString;
 
 class CarouselController extends Controller
