@@ -2,13 +2,11 @@
 
 namespace App\Providers;
 
+use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\ServiceProvider;
-use Illuminate\Pagination\Paginator;
-use Illuminate\Support\Facades\Schema;  // <== Mandatory for production use
 
 class AppServiceProvider extends ServiceProvider
 {
-
     /**
      * Register any application services.
      *
@@ -26,7 +24,6 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        //Paginator::useBootstrap();
         Schema::defaultStringLength(191);  // <== Mandatory for production use
     }
 }
