@@ -2,11 +2,10 @@
 
 @section('content')
 
-
     <div class="row">
         <div class="col-1 ml-n5">
             <a href="{{ route('events.show', $event) }}"><i class="fa fa-4x fa-arrow-circle-left return fa-return growIcon"
-                    aria-hidden="true"></i></a>
+                                                            aria-hidden="true"></i></a>
         </div>
         <div class="col-11 ml-n2">
             <h1>S'insrire à
@@ -22,9 +21,9 @@
                 @csrf
                 <div class="form-group">
                     <label for="tournamentName">Veuillez sélectionner le rôle pour cet événement (notez qu'un seul rôle
-                        pourra être choisi).</label>
+                                                pourra être choisi).</label>
                     <select class="form-control" name="engagement">
-                        @foreach (\App\Role::availableForEngagement()->get() as $engagement)
+                        @foreach (\App\Models\Role::availableForEngagement()->get() as $engagement)
                             <option value="{{ $engagement->slug }}">{{ $engagement->name }}</option>
                         @endforeach
                     </select>
