@@ -3,10 +3,10 @@
 @section('content')
 
     <?php
-        $mainArray = array ();
-        foreach ($pool->teams as $team){
-            array_push($mainArray,[$team->id,$team->name]);
-        }
+    $mainArray = array();
+    foreach ($pool->teams as $team) {
+        array_push($mainArray, [$team->id, $team->name]);
+    }
 
     ?>
     <div class="container">
@@ -31,32 +31,32 @@
                 <button type="submit" class="btn btn-main" data-toggle="modal" data-target="#stagePoolModal">Passer à l'étape suivante</button>
             @endif
             @if ($pool->stage > 1 && count($pool->contenders) < $pool->poolSize)
-            <h2>Définition des équipes</h2>
-            <table>
-                <tr>
-                    <th>Issu de la poule</th>
-                    <th>Classé</th>
-                </tr>
-                <tr>
-                    <form action="{{ route('pools.contenders.store', $pool->id) }}" method="post">
-                        @csrf
-                        <td>
-                            <select name="pool_from_id" id="pool_from_id">
-                                @foreach ($poolsInPreviousStage as $poolInPreviousStage)
-                                <option value="{{ $poolInPreviousStage->id }}">{{ $poolInPreviousStage->poolName }}</option>
-                                @endforeach
-                            </select>
-                        </td>
-                        <td>
-                            <input type="number" name="rank_in_pool" id="rank_in_pool" min="1">
+                <h2>Définition des équipes</h2>
+                <table>
+                    <tr>
+                        <th>Issu de la poule</th>
+                        <th>Classé</th>
+                    </tr>
+                    <tr>
+                        <form action="{{ route('pools.contenders.store', $pool->id) }}" method="post">
+                            @csrf
+                            <td>
+                                <select name="pool_from_id" id="pool_from_id">
+                                    @foreach ($poolsInPreviousStage as $poolInPreviousStage)
+                                        <option value="{{ $poolInPreviousStage->id }}">{{ $poolInPreviousStage->poolName }}</option>
+                                    @endforeach
+                                </select>
+                            </td>
+                            <td>
+                                <input type="number" name="rank_in_pool" id="rank_in_pool" min="1">
 
-                        </td>
-                        <td>
-                            <button type="submit" class="btn btn-main">+</button>
-                        </td>
-                    </form>
-                </tr>
-            </table>
+                            </td>
+                            <td>
+                                <button type="submit" class="btn btn-main">+</button>
+                            </td>
+                        </form>
+                    </tr>
+                </table>
             @endif
             <h4>Date : {{$tournament->start_date->format('d.m.Y')}}</h4>
             <div class="row justify-content-center">
@@ -73,14 +73,14 @@
                                     @elseif(isset($game->contender1->team->name))
                                         <td class="contender1">{{ $game->contender1->team->name }}</td>
                                     @else
-                                    <td>
-                                    <select id="inputState" class="form-control">
-                                        <option selected>Choisir team</option>
-                                        @foreach($mainArray as $value)
-                                            <option>{{ $value[1] }}</option>
-                                        @endforeach
-                                    </select>
-                                    </td>
+                                        <td>
+                                            <select id="inputState" class="form-control">
+                                                <option selected>Choisir team</option>
+                                                @foreach($mainArray as $value)
+                                                    <option>{{ $value[1] }}</option>
+                                                @endforeach
+                                            </select>
+                                        </td>
                                     @endif
                                     <td class="separator sepTime">{{Carbon\Carbon::parse($game->start_time)->format('H:i')}}</td>
                                     <td class="score2">{{ $game->court->name }}</td>
@@ -91,14 +91,14 @@
                                         <td class="contender2">{{ $game->contender2->team->name }}</td>
 
                                     @else
-                                    <td>
-                                    <select id="inputState" class="form-control">
-                                        <option selected>Choisir team</option>
-                                        @foreach($mainArray as $value)
-                                            <option>{{ $value[1] }}</option>
-                                        @endforeach
-                                    </select>
-                                    </td>
+                                        <td>
+                                            <select id="inputState" class="form-control">
+                                                <option selected>Choisir team</option>
+                                                @foreach($mainArray as $value)
+                                                    <option>{{ $value[1] }}</option>
+                                                @endforeach
+                                            </select>
+                                        </td>
                                     @endif
 
                                     @if($pool->isEditable())
@@ -106,11 +106,11 @@
                                     @endif
                                 @else
 
-                                     <!-- teams - no score -->
+                                    <!-- teams - no score -->
                                     @if(!isset($game->score_contender1) || !isset($game->score_contender2))
                                         <td class="separator sepTime ">{{Carbon\Carbon::parse($game->start_time)->format('H:i')}}</td>
                                         <td class="contender1 ">{{$game->contender1->team->name}}</td>
-                                        <td class="separator"> - </td>
+                                        <td class="separator"> -</td>
                                         <td class="contender2">{{$game->contender2->team->name}}</td>
                                         <td class="score2 ">{{ $game->court->name }}</td>
                                         @if($pool->isEditable())
@@ -119,25 +119,25 @@
 
                                     @else
                                         <!--teams and score -->
-                                        <tr style="background-color: #DCDCDC;">
-                                            <td class="contender1">{{$game->contender1->team->name}}</td>
-                                            <td class="score1">{{$game->score_contender1}}</td>
-                                            <td class="separator"> - </td>
-                                            <td class="score2">{{$game->score_contender2}}</td>
-                                            <td class="contender2">{{$game->contender2->team->name}}</td>
-                                        </tr>
-                                        @if($pool->isEditable())
-                                            <td class="action"><i class="fa fa-lg fa-trophy editScore" aria-hidden="true"></i></td>
-                                        @endif
-                                    @endif
-                                @endif
+                            <tr style="background-color: #DCDCDC;">
+                                <td class="contender1">{{$game->contender1->team->name}}</td>
+                                <td class="score1">{{$game->score_contender1}}</td>
+                                <td class="separator"> -</td>
+                                <td class="score2">{{$game->score_contender2}}</td>
+                                <td class="contender2">{{$game->contender2->team->name}}</td>
                             </tr>
-                        @endforeach
-                    @else
+                            @if($pool->isEditable())
+                                <td class="action"><i class="fa fa-lg fa-trophy editScore" aria-hidden="true"></i></td>
+                                @endif
+                                @endif
+                                @endif
+                                </tr>
+                                @endforeach
+                            @else
 
-                        Aucun match pour l'instant ...
+                                Aucun match pour l'instant ...
 
-                    @endif
+                            @endif
                     </tbody>
                 </table>
             </div>
@@ -171,67 +171,65 @@
                     @endfor
                     </tbody>
                 </table>
-            @elseif(\App\Contender::isAllEmpty($contenders))
+            @elseif(\App\Models\Contender::isAllEmpty($contenders))
                 <h2>Liste des participants</h2>
                 <table class="table">
 
 
-
-
-
-                @for($i = 0;$i < $pool->poolSize; $i++ )
-                    <tr>
-                        <td>
-                            <h6 style="color: black">
-                                @if(isset($mainArray[$i][1]))
-                                    {{ $mainArray[$i][1] }}
-                                @else
-                                    !Vide!
-                                @endif
-                            </h6>
+                    @for($i = 0;$i < $pool->poolSize; $i++ )
+                        <tr>
+                            <td>
+                                <h6 style="color: black">
+                                    @if(isset($mainArray[$i][1]))
+                                        {{ $mainArray[$i][1] }}
+                                    @else
+                                        !Vide!
+                                    @endif
+                                </h6>
                             </td>
-                        <td>
-                        <td>
-                        @if((isset($pool->id)) && (isset($mainArray[$i][0])))
-                            <form action="{{ route('pools.contenders.destroy', [$mainArray[$i][0], $pool->id]) }}" method="POST">
-                                @csrf
-                                <input type="hidden" name="_method" value="DELETE">
-                                <button type="submit" class="btn btn-danger">
-                                    <i class="fa fa-trash fa-lg" aria-hidden="true"></i>
-                                </button>
-                            </form>
-                        @endif
-                        </td>
-                        </td>
-                    </tr>
-                @endfor
+                            <td>
+                            <td>
+                                @if((isset($pool->id)) && (isset($mainArray[$i][0])))
+                                    <form action="{{ route('pools.contenders.destroy', [$mainArray[$i][0], $pool->id]) }}" method="POST">
+                                        @csrf
+                                        <input type="hidden" name="_method" value="DELETE">
+                                        <button type="submit" class="btn btn-danger">
+                                            <i class="fa fa-trash fa-lg" aria-hidden="true"></i>
+                                        </button>
+                                    </form>
+                                @endif
+                            </td>
+                            </td>
+                        </tr>
+                    @endfor
                 </table>
             @else
                 <h2>Liste des participants</h2>
                 @foreach($contenders as $contender)
-                    <h6 style="color: black" value="{{$contender[' pool_from_id']}}">{{$contender->rank_in_pool .($contender->rank_in_pool == 1 ? "er " : 'ème ') . "de "  . $contender->fromPool->poolName}}</h6>
+                    <h6 style="color: black"
+                        value="{{$contender[' pool_from_id']}}">{{$contender->rank_in_pool .($contender->rank_in_pool == 1 ? "er " : 'ème ') . "de "  . $contender->fromPool->poolName}}</h6>
                 @endforeach
             @endif
             @if($pool->stage == 1 && count($teamsNotInAPool) > 0)
-            <form action="{{ route('pools.contenders.store', $pool->id) }}" method="post">
-                @csrf
-                <table class="table">
+                <form action="{{ route('pools.contenders.store', $pool->id) }}" method="post">
+                    @csrf
+                    <table class="table">
 
-                    <tr>
-                        <td>
-                            <select id="teams" name="team_id" class="form-control">
-                                @foreach ($teamsNotInAPool as $team)
-                                    <option value="{{ $team->id }}">{{ $team->name }}</option>
-                                @endforeach
-                            </select>
+                        <tr>
+                            <td>
+                                <select id="teams" name="team_id" class="form-control">
+                                    @foreach ($teamsNotInAPool as $team)
+                                        <option value="{{ $team->id }}">{{ $team->name }}</option>
+                                    @endforeach
+                                </select>
                             </td>
                             <td>
-                            <button type="submit" class="btn btn-main">Ajouter</button>
-                    </td>
-                    </tr>
-                </table>
+                                <button type="submit" class="btn btn-main">Ajouter</button>
+                            </td>
+                        </tr>
+                    </table>
 
-            </form>
+                </form>
             @endif
         </div>
     </div>
